Fixed bug with transaction fee not showing

+++ app/src/main/java/com/dcrandroid/data/Transaction.java
@@ -55,7 +55,7 @@ public class Transaction implements Serializable{
         this.amount = amount;
     }
 
-    public float getTransactionFee() {
+    public long getTransactionFee() {
         return TransactionFee;
     }
